fix(infra): share parameter parsing with gcp
