+++ internal/client/infra/deploy.go
@@ -14,8 +14,8 @@ import (
 )
 
 const (
-	// parameterSplitParts is the expected number of parts when splitting a KEY=VALUE parameter.
-	parameterSplitParts = 2
+        // parameterSplitParts is the expected number of parts when splitting a KEY=VALUE parameter.
+        parameterSplitParts = 2
 )
 
 // NewDeployer creates a Deployer for the specified provider.
@@ -82,17 +82,3 @@ func resolveAWSTemplate(template, version, region string) (*core.TemplateSource,
 	return &core.TemplateSource{Body: string(content)}, nil
 }
 
-// ParseParameters parses KEY=VALUE parameter strings.
-func ParseParameters(params []string) (map[string]string, error) {
-	result := make(map[string]string)
-
-	for _, param := range params {
-		parts := strings.SplitN(param, "=", parameterSplitParts)
-		if len(parts) != parameterSplitParts {
-			return nil, fmt.Errorf("invalid parameter format: %s (expected KEY=VALUE)", param)
-		}
-		result[parts[0]] = parts[1]
-	}
-
-	return result, nil
-}
